requirements.txt
streamlit
sentence-transformers
numpy
pandas
scikit-learn
plotly
requests
zhconv
dashscope
beautifulsoup4
python-dotenv
jieba